Made the following speed optimizations: 	- Added a disable_checks flag to disable the rules engine from doing extra input validation 	- Made all numeric comparisons use float not Decimal 	- Defaulted to NOT assert_type_for_arguments as that can slow down processing

+++ tests/operators/test_operators.py
@@ -86,10 +86,10 @@ class NumericOperatorTests(TestCase):
         ten_var_int = NumericType(ten_int)
         ten_var_float = NumericType(ten_float)
         ten_var_long = NumericType(ten_long)
-        self.assertTrue(isinstance(ten_var_dec.value, Decimal))
-        self.assertTrue(isinstance(ten_var_int.value, Decimal))
-        self.assertTrue(isinstance(ten_var_float.value, Decimal))
-        self.assertTrue(isinstance(ten_var_long.value, Decimal))
+        self.assertTrue(isinstance(ten_var_dec.value, float))
+        self.assertTrue(isinstance(ten_var_int.value, int))
+        self.assertTrue(isinstance(ten_var_float.value, float))
+        self.assertTrue(isinstance(ten_var_long.value, int))
 
     def test_numeric_equal_to(self):
         self.assertTrue(NumericType(10).equal_to(10))
